requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0

pydantic-settings==2.12.0
python-dotenv==1.2.1

langchain-core==1.2.5
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0

sentence-transformers==5.2.0
faiss-cpu==1.13.2

openai==2.14.0